Reuse CornerLabelRenderer on all grids
- replace all instances of cornerbuttonmixin with cornerlabelrenderer
- some cleanup
